requirements.txt
flask
Flask-SQLAlchemy
argon2-cffi